Create generate_results.m
Runs the all the functions for the project on an image, and shows the results. After reading in the image, it calls reduceWidth or reduceHeight as many times as needed to reduce the width/height by the requested amount, but only sets the display flag to true in the first iteration. You should also display, in a 1x3 subplot: (a) the original image, (b) the final content-aware-reduced image, and (c) the content- agnostic reduction using a simple imresize.