Create reduceWidth.m
Inputs:
The function should take in a nrows-x-ncols-x-3 matrix im of class uint8. If the display_flag is set to true, the function should display, in a 1x3 subplot, the following: (a) the energy image, (b) the cumulative energy image, (c) the optimal seam (using displaySeam).

Outputs:
The output must be a 3D matrix reducedColorImage, of class uint8, which is the same as the input image but with its width reduced by one pixel.